Delete ModelStub and return to original
